Move mfa expiry logic into concern
