+++ app/controllers/passwords_controller.rb
@@ -1,4 +1,6 @@
 class PasswordsController < Clearance::PasswordsController
+  include MfaExpiryMethods
+
   before_action :validate_confirmation_token, only: %i[edit mfa_edit webauthn_edit]
 
   def edit
@@ -6,7 +8,7 @@ class PasswordsController < Clearance::PasswordsController
       setup_mfa_authentication
       setup_webauthn_authentication
 
-      session[:mfa_expires_at] = 15.minutes.from_now.to_s
+      create_new_mfa_expiry
 
       render template: "multifactor_auths/mfa_prompt"
     else
@@ -38,7 +40,7 @@ class PasswordsController < Clearance::PasswordsController
       login_failure(t("multifactor_auths.incorrect_otp"))
     end
   ensure
-    session.delete(:mfa_expires_at)
+    delete_mfa_expiry_session
   end
 
   def webauthn_edit
@@ -69,7 +71,7 @@ class PasswordsController < Clearance::PasswordsController
   rescue WebAuthn::Error => e
     login_failure(e.message)
   ensure
-    session.delete(:mfa_expires_at)
+    delete_mfa_expiry_session
   end
 
   private
@@ -116,10 +118,5 @@ class PasswordsController < Clearance::PasswordsController
   def login_failure(message)
     flash.now.alert = message
     render template: "multifactor_auths/mfa_prompt", status: :unauthorized
-  end
-
-  def session_active?
-    return false if session[:mfa_expires_at].nil?
-    session[:mfa_expires_at] > Time.current
   end
 end
